# packages/frontend/backend_openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payments": {
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "summary": "Create a new payment",
        "tags": [
          "payments"
        ]
      },
      "get": {
        "operationId": "PaymentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of payments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all payments",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/student/{studentId}": {
      "get": {
        "operationId": "PaymentsController_findByStudent",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payments for student",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          }
        },
        "summary": "Get payments by student id",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "summary": "Get payment by id",
        "tags": [
          "payments"
        ]
      },
      "put": {
        "operationId": "PaymentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "summary": "Update payment by id",
        "tags": [
          "payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment deleted"
          }
        },
        "summary": "Delete payment by id",
        "tags": [
          "payments"
        ]
      }
    },
    "/groups": {
      "get": {
        "operationId": "GroupsController_getAll",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all groups or filtered by active status"
          }
        },
        "summary": "Get all groups",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created"
          }
        },
        "summary": "Create new group",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}": {
      "get": {
        "operationId": "GroupsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns group"
          }
        },
        "summary": "Get group by ID",
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "operationId": "GroupsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group updated"
          }
        },
        "summary": "Update group",
        "tags": [
          "groups"
        ]
      },
      "delete": {
        "operationId": "GroupsController_deleteGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group deleted"
          }
        },
        "summary": "Delete group",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/batch-upsert": {
      "post": {
        "operationId": "GroupsController_batchUpsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpsertGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Groups created/updated. Returns count and the processed groups."
          }
        },
        "summary": "Batch create or update groups by name",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/debits": {
      "post": {
        "operationId": "DebitsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDebitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Debit created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Debit"
                }
              }
            }
          }
        },
        "summary": "Create a new debit",
        "tags": [
          "debits"
        ]
      },
      "get": {
        "operationId": "DebitsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of debits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Debit"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all debits",
        "tags": [
          "debits"
        ]
      }
    },
    "/debits/{id}": {
      "get": {
        "operationId": "DebitsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debit found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Debit"
                }
              }
            }
          }
        },
        "summary": "Get debit by id",
        "tags": [
          "debits"
        ]
      },
      "put": {
        "operationId": "DebitsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDebitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Debit updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Debit"
                }
              }
            }
          }
        },
        "summary": "Update debit by id",
        "tags": [
          "debits"
        ]
      },
      "delete": {
        "operationId": "DebitsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debit deleted"
          }
        },
        "summary": "Delete debit by id",
        "tags": [
          "debits"
        ]
      }
    },
    "/group-templates": {
      "get": {
        "operationId": "GroupTemplatesController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all group templates"
          }
        },
        "summary": "Get all group templates",
        "tags": [
          "group-templates"
        ]
      },
      "post": {
        "operationId": "GroupTemplatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group template created"
          }
        },
        "summary": "Create new group template",
        "tags": [
          "group-templates"
        ]
      }
    },
    "/group-templates/{id}": {
      "get": {
        "operationId": "GroupTemplatesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns group template"
          }
        },
        "summary": "Get group template by ID",
        "tags": [
          "group-templates"
        ]
      },
      "patch": {
        "operationId": "GroupTemplatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group template updated"
          }
        },
        "summary": "Update group template",
        "tags": [
          "group-templates"
        ]
      },
      "delete": {
        "operationId": "GroupTemplatesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group template deleted"
          }
        },
        "summary": "Delete group template",
        "tags": [
          "group-templates"
        ]
      }
    },
...
  },
  "info": {
    "title": "KRUSANT API",
    "description": "API documentation for KRUSANT backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "date",
            "example": "2026-01-15"
          },
          "amount": {
            "type": "number",
            "example": 100.5
          },
          "comment": {
            "type": "string",
            "example": "Some comment"
          },
          "proofType": {
            "type": "string",
            "enum": [
              "receipt",
              "invoice"
            ],
            "example": "receipt"
          },
          "fiscalized": {
            "type": "boolean",
            "example": false
          },
          "invoiceId": {
            "type": "number",
            "example": 123
          },
          "studentId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "date",
          "amount",
          "proofType",
          "fiscalized",
          "studentId"
        ]
      },
      "Payment": {
        "type": "object",
        "properties": {}
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "date",
            "example": "2026-01-15"
          },
          "amount": {
            "type": "number",
            "example": 100.5
          },
          "comment": {
            "type": "string",
            "example": "Some comment"
          },
          "proofType": {
            "type": "string",
            "enum": [
              "receipt",
              "invoice"
            ],
            "example": "receipt"
          },
          "fiscalized": {
            "type": "boolean",
            "example": false
          },
          "invoiceId": {
            "type": "number",
            "example": 123
          },
          "studentId": {
            "type": "number",
            "example": 1
          }
        }
      },
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Group name",
            "example": "Group A"
          },
          "isActive": {
            "type": "boolean",
            "description": "Group active status",
            "example": true,
            "default": true
          },
          "studentIds": {
            "description": "Array of student IDs",
            "example": [
              1,
              2,
              3
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "classIds": {
            "description": "Array of class IDs",
            "example": [
              1,
              2,
              3
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minStartDate": {
            "description": "Minimum start date constraint",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateBoundaryDto"
              }
            ]
          },
          "maxEndDate": {
            "description": "Maximum end date constraint",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateBoundaryDto"
              }
            ]
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          },
          "cost": {
            "type": "number",
            "description": "Total course cost",
            "example": 3325
          },
          "unitCost": {
            "type": "number",
            "description": "Cost per unit",
            "example": 50
          },
          "comment": {
            "type": "string",
            "description": "Additional comments",
            "example": "",
            "default": ""
          },
          "colorHex": {
            "type": "string",
            "description": "Group color hex for visual identification",
            "example": "#4CAF50"
          },
          "startHour": {
            "type": "string",
            "description": "Start hour as time (HH:mm or HH:mm:ss)",
            "example": "09:00"
          },
          "lessonLength": {
            "type": "string",
            "description": "Lesson length as time (HH:mm or HH:mm:ss)",
            "example": "01:00"
          },
          "roomId": {
            "type": "number",
            "description": "Room ID",
            "example": 1
          }
        },
        "required": [
          "name",
          "teacherId",
          "cost",
          "unitCost"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Group name",
            "example": "Group A"
          },
          "isActive": {
            "type": "boolean",
            "description": "Group active status",
            "example": true,
            "default": true
          },
          "studentIds": {
            "description": "Array of student IDs",
            "example": [
              1,
              2,
              3
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "classIds": {
            "description": "Array of class IDs",
            "example": [
              1,
              2,
              3
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minStartDate": {
            "description": "Minimum start date constraint",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateBoundaryDto"
              }
            ]
          },
          "maxEndDate": {
            "description": "Maximum end date constraint",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateBoundaryDto"
              }
            ]
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          },
          "cost": {
            "type": "number",
            "description": "Total course cost",
            "example": 3325
          },
          "unitCost": {
            "type": "number",
            "description": "Cost per unit",
            "example": 50
          },
          "comment": {
            "type": "string",
            "description": "Additional comments",
            "example": "",
            "default": ""
          },
          "colorHex": {
            "type": "string",
            "description": "Group color hex for visual identification",
            "example": "#4CAF50"
          },
          "startHour": {
            "type": "string",
            "description": "Start hour as time (HH:mm or HH:mm:ss)",
            "example": "09:00"
          },
          "lessonLength": {
            "type": "string",
            "description": "Lesson length as time (HH:mm or HH:mm:ss)",
            "example": "01:00"
          },
          "roomId": {
            "type": "number",
            "description": "Room ID",
            "example": 1
          }
        }
      },
      "BatchUpsertGroupDto": {
        "type": "object",
        "properties": {
          "groups": {
            "description": "Array of groups to create or update",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateGroupDto"
            }
          }
        },
        "required": [
          "groups"
        ]
      },
...
      "CreateDebitDto": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "date",
            "example": "2026-01-15"
          },
          "amount": {
            "type": "number",
            "example": 100.5
          },
          "comment": {
            "type": "string",
            "example": "Some comment"
          },
          "entitlement": {
            "type": "string",
            "example": "Scholarship"
          },
          "studentId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "dueDate",
          "amount",
          "studentId"
        ]
      },
      "Debit": {
        "type": "object",
        "properties": {}
      },
      "UpdateDebitDto": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "date",
            "example": "2026-01-15"
          },
          "amount": {
            "type": "number",
            "example": 100.5
          },
          "comment": {
            "type": "string",
            "example": "Some comment"
          },
          "entitlement": {
            "type": "string",
            "example": "Scholarship"
          },
          "studentId": {
            "type": "number",
            "example": 1
          }
        }
      },
      "CreateGroupTemplateDto": {
        "type": "object",
        "properties": {
          "templateName": {
            "type": "string",
            "description": "Template name",
            "example": "Evening Regular"
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status",
            "example": true,
            "default": true
          },
          "studentIds": {
            "description": "Student IDs",
            "example": [
              1,
              2
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "classIds": {
            "description": "Class IDs",
            "example": [
              1,
              2
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          },
          "cost": {
            "type": "number",
            "description": "Total course cost",
            "example": 3325
          },
          "unitCost": {
            "type": "number",
            "description": "Cost per unit",
            "example": 50
          },
          "comment": {
            "type": "string",
            "description": "Additional comments",
            "example": ""
          },
          "minStartDate": {
            "description": "Minimum start date constraint",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateBoundaryDto"
              }
            ]
          },
          "maxEndDate": {
            "description": "Maximum end date constraint",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateBoundaryDto"
              }
            ]
          },
          "colorHex": {
            "type": "string",
            "description": "Template color hex for visual identification",
            "example": "#4CAF50"
          },
          "startHour": {
            "type": "string",
            "description": "Default start time (HH:mm or HH:mm:ss)",
            "example": "09:00",
            "format": "time"
          },
          "lessonLength": {
            "type": "string",
            "description": "Lesson duration as time (HH:mm or HH:mm:ss)",
            "example": "01:00",
            "format": "time"
          },
          "roomId": {
            "type": "number",
            "description": "Room ID",
            "example": 1
          },
          "courseId": {
            "type": "number",
            "description": "Course ID",
            "example": 1
          }
        },
        "required": [
          "templateName"
        ]
      },
...
    }
  }
}